% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalCDF.R
\name{ConditionalCDF}
\alias{ConditionalCDF}
\title{Estimate conditional cumulative distribution function}
\usage{
ConditionalCDF(Y_eval, X_eval, Y_train, X_train, h_y, h_x)
}
\arguments{
\item{Y_eval}{A numeric vector of Y values where the CDF should be evaluated.}

\item{X_eval}{A numeric matrix of X values where the CDF should be evaluated.}

\item{Y_train}{A numeric vector of training Y values.}

\item{X_train}{A numeric matrix of training X values.}

\item{h_y}{A numeric scalar representing the bandwidth for the response variable Y.}

\item{h_x}{A numeric vector representing the bandwidths for the covariates X.}
}
\value{
A numeric vector of estimated CDF probabilities.
}
\description{
Computes a doubly smoothed Nadaraya-Watson estimator for the conditional
cumulative distribution function
}
